requirements.txt
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
colorama==0.4.4
et-xmlfile==1.1.0
Flask==2.0.1
google-api-core==2.0.1
google-auth==2.1.0
google-cloud-core==2.0.0
google-cloud-storage==1.42.2
google-crc32c==1.2.0
google-resumable-media==2.0.3
googleapis-common-protos==1.53.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
numpy==1.21.2
openpyxl==3.0.7
pandas==1.3.2
Pillow==8.3.2
protobuf==3.18.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyPDF2==1.26.0
python-barcode==0.13.1
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
requests==2.26.0
rsa==4.7.2
six==1.16.0
urllib3==1.26.6
xlrd==2.0.1
Werkzeug==2.0.1